reverted cam structure
